[LP-23] Refactor:
Modificações:

Alteração dos nomes dos arquivos e implementação do @Data
