+++ src/main/java/br/edu/ifpb/pweb2/lumicash/entity/Correntista.java
@@ -94,4 +94,5 @@ public class Correntista {
     public void setContas(List<Conta> contas) {
         this.contas = contas;
     }
+
 }
